+++ Controller/AdLostController.php
@@ -46,115 +46,41 @@ class AdLostController {
      * @param $ad
      */
     public function addAd($ad, $files) {
-        if (isset($ad['animal'], $ad['name'], $ad['sex'], $ad['size'], $ad['fur'], $ad['color'], $ad['dress'], $ad['race'],
-        $ad['number'], $ad['description'], $ad['date_lost'], $ad['date'], $ad['city'], $files['picture'], $ad['user_fk'])) {
+        if (isset($_SESSION["id"])) {
+            if (isset($ad['animal'], $ad['name'], $ad['sex'], $ad['size'], $ad['fur'], $ad['color'], $ad['dress'], $ad['race'],
+                $ad['number'], $ad['description'], $ad['date_lost'], $ad['date'], $ad['city'], $files['picture'], $ad['user_fk'])) {
 
-            $userManager = new UserManager();
-            $adlostManager = new AdLostManager();
+                $userManager = new UserManager();
+                $adlostManager = new AdLostManager();
 
-            $animal = htmlentities($ad['animal']);
-            $name = htmlentities(ucfirst($ad['name']));
-            $sex = htmlentities($ad['sex']);
-            $size = htmlentities($ad['size']);
-            $fur = htmlentities($ad['fur']);
-            $dress = htmlentities($ad['dress']);
-            $race = htmlentities(ucfirst($ad['race']));
-            $number = htmlentities($ad['number']);
-            $description = htmlentities(ucfirst($ad['description']));
-            $date_lost = htmlentities($ad['date_lost']);
-            $date = htmlentities($ad['date']);
-            $city = htmlentities($ad['city']);
-            $user_fk = intval($ad['user_fk']);
+                $animal = $ad['animal'];
+                $name = htmlentities(ucfirst($ad['name']));
+                $sex = $ad['sex'];
+                $size = $ad['size'];
+                $fur = $ad['fur'];
+                $dress = $ad['dress'];
+                $race = htmlentities(ucfirst($ad['race']));
+                $number = htmlentities(strtoupper($ad['number']));
+                $description = htmlentities(ucfirst($ad['description']));
+                $date_lost = $ad['date_lost'];
+                $date = $ad['date'];
+                $city = $ad['city'];
+                $user_fk = intval($ad['user_fk']);
 
-            if (count($ad['color']) === 1) {
-                $color = $ad['color'][0];
-            }
-            elseif (count($ad['color']) === 2) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1];
-            }
-            elseif (count($ad['color']) === 3) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2];
-            }
-            elseif (count($ad['color']) === 4) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3];
-            }
-            elseif (count($ad['color']) === 5) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4];
-            }
-            else {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " .$ad['color'][4] . ", " . $ad['color'][5];
-            }
-
-            if (in_array($files['picture']['type'], ['image/jpg', 'image/jpeg', 'image/png', ".jpg"])) {
-                $maxSize = 2 * 1024 * 1024; // = 2 Mo
-
-                if ($files['picture']['size'] <= $maxSize) {
-                    $tmpName = $files['picture']['tmp_name'];
-                    $namePicture = getRandomName($files['picture']['name']);
-
-                    move_uploaded_file($tmpName, "./assets/img/adLost/".$namePicture);
-
-                    $user_fk = $userManager->getUser($user_fk);
-                    if($user_fk->getId()) {
-                        $ad = new AdLost(null, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $namePicture, $user_fk);
-                        $adlostManager->add($ad);
-                        header("Location: ../index.php?controller=adlost&action=view&success=3");
-                    }
+                if (count($ad['color']) === 1) {
+                    $color = $ad['color'][0];
+                } elseif (count($ad['color']) === 2) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1];
+                } elseif (count($ad['color']) === 3) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2];
+                } elseif (count($ad['color']) === 4) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3];
+                } elseif (count($ad['color']) === 5) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4];
+                } else {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4] . ", " . $ad['color'][5];
                 }
-                else {
-                    header("Location: ../index.php?controller=adlost&action=new&error=1");
-                }
-            }
-            else {
-                header("Location: ../index.php?controller=adlost&action=new&error=0");
-            }
-        }
-        $this->return("create/addLostView", "Anim'Nord : Ajouter une annonce de chiens et chats perdus");
-    }
 
-    public function updateAd($ad, $files) {
-        if (isset($ad['id'], $ad['animal'], $ad['name'], $ad['sex'], $ad['size'], $ad['fur'], $ad['color'], $ad['dress'], $ad['race'],
-            $ad['number'], $ad['description'], $ad['date_lost'], $ad['date'], $ad['city'], $ad['picture2'], $ad['user_fk'])) {
-
-            $userManager = new UserManager();
-            $adlostManager = new AdLostManager();
-
-            $id = intval($ad['id']);
-            $animal = htmlentities($ad['animal']);
-            $name = htmlentities(ucfirst($ad['name']));
-            $sex = htmlentities($ad['sex']);
-            $size = htmlentities($ad['size']);
-            $fur = htmlentities($ad['fur']);
-            $dress = htmlentities($ad['dress']);
-            $race = htmlentities(ucfirst($ad['race']));
-            $number = htmlentities($ad['number']);
-            $description = htmlentities(ucfirst($ad['description']));
-            $date_lost = htmlentities($ad['date_lost']);
-            $date = htmlentities($ad['date']);
-            $city = htmlentities($ad['city']);
-            $picture = htmlentities($ad['picture2']);
-            $user_fk = intval($ad['user_fk']);
-
-            if (count($ad['color']) === 1) {
-                $color = $ad['color'][0];
-            }
-            elseif (count($ad['color']) === 2) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1];
-            }
-            elseif (count($ad['color']) === 3) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2];
-            }
-            elseif (count($ad['color']) === 4) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3];
-            }
-            elseif (count($ad['color']) === 5) {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4];
-            }
-            else {
-                $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " .$ad['color'][4] . ", " . $ad['color'][5];
-            }
-
-            if (isset($files['picture'])) {
                 if (in_array($files['picture']['type'], ['image/jpg', 'image/jpeg', 'image/png', ".jpg"])) {
                     $maxSize = 2 * 1024 * 1024; // = 2 Mo
 
@@ -162,35 +88,103 @@ class AdLostController {
                         $tmpName = $files['picture']['tmp_name'];
                         $namePicture = getRandomName($files['picture']['name']);
 
-                        move_uploaded_file($tmpName, "./assets/img/adLost/".$namePicture);
-                        unlink("./assets/img/adLost/" . $picture);
+                        move_uploaded_file($tmpName, "./assets/img/adLost/" . $namePicture);
 
                         $user_fk = $userManager->getUser($user_fk);
                         if ($user_fk->getId()) {
-                            $ad = new AdLost($id, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $namePicture, $user_fk);
-                            $adlostManager->update($ad);
-                            header("Location: ../index.php?controller=adlost&action=view&success=1");
+                            $ad = new AdLost(null, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $namePicture, $user_fk);
+                            $adlostManager->add($ad);
+                            header("Location: ../index.php?controller=adlost&action=view&success=3");
                         }
+                    } else {
+                        header("Location: ../index.php?controller=adlost&action=new&error=1");
                     }
-                    else {
-                        header("Location: ../index.php?controller=adlost&action=update&id=$id&error=1");
-                    }
+                } else {
+                    header("Location: ../index.php?controller=adlost&action=new&error=0");
                 }
-                else {
-                    header("Location: ../index.php?controller=adlost&action=update&id=$id&error=0");
+            }
+            $this->return("create/addLostView", "Anim'Nord : Ajouter une annonce de chiens et chats perdus");
+        }
+    }
+
+    /**
+     * update a ad
+     * @param $ad
+     * @param $files
+     */
+    public function updateAd($ad, $files) {
+        if (isset($_SESSION["id"])) {
+            if (isset($ad['id'], $ad['animal'], $ad['name'], $ad['sex'], $ad['size'], $ad['fur'], $ad['color'], $ad['dress'], $ad['race'],
+                $ad['number'], $ad['description'], $ad['date_lost'], $ad['date'], $ad['city'], $ad['picture2'], $ad['user_fk'])) {
+
+                $userManager = new UserManager();
+                $adlostManager = new AdLostManager();
+
+                $id = intval($ad['id']);
+                $animal = $ad['animal'];
+                $name = htmlentities(ucfirst($ad['name']));
+                $sex = $ad['sex'];
+                $size = $ad['size'];
+                $fur = $ad['fur'];
+                $dress = $ad['dress'];
+                $race = htmlentities(ucfirst($ad['race']));
+                $number = htmlentities(strtoupper($ad['number']));
+                $description = htmlentities(ucfirst($ad['description']));
+                $date_lost = $ad['date_lost'];
+                $date = $ad['date'];
+                $city = $ad['city'];
+                $picture = htmlentities($ad['picture2']);
+                $user_fk = intval($ad['user_fk']);
+
+                if (count($ad['color']) === 1) {
+                    $color = $ad['color'][0];
+                } elseif (count($ad['color']) === 2) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1];
+                } elseif (count($ad['color']) === 3) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2];
+                } elseif (count($ad['color']) === 4) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3];
+                } elseif (count($ad['color']) === 5) {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4];
+                } else {
+                    $color = $ad['color'][0] . ", " . $ad['color'][1] . ", " . $ad['color'][2] . ", " . $ad['color'][3] . ", " . $ad['color'][4] . ", " . $ad['color'][5];
                 }
 
-            }
-            else {
-                $user_fk = $userManager->getUser($user_fk);
-                if ($user_fk->getId()) {
-                    $ad = new AdLost($id, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $picture, $user_fk);
-                    $adlostManager->update($ad);
-                    header("Location: ../index.php?controller=adlost&action=view&success=1");
+                if (isset($ad['picture'])) {
+                    if (in_array($files['picture']['type'], ['image/jpg', 'image/jpeg', 'image/png', ".jpg"])) {
+                        $maxSize = 2 * 1024 * 1024; // = 2 Mo
+
+                        if ($files['picture']['size'] <= $maxSize) {
+                            $tmpName = $files['picture']['tmp_name'];
+                            $namePicture = getRandomName($files['picture']['name']);
+
+                            move_uploaded_file($tmpName, "./assets/img/adLost/" . $namePicture);
+                            unlink("./assets/img/adLost/" . $picture);
+
+                            $user_fk = $userManager->getUser($user_fk);
+                            if ($user_fk->getId()) {
+                                $ad = new AdLost($id, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $namePicture, $user_fk);
+                                $adlostManager->update($ad);
+                                header("Location: ../index.php?controller=adlost&action=view&success=1");
+                            }
+                        } else {
+                            header("Location: ../index.php?controller=adlost&action=update&id=$id&error=1");
+                        }
+                    } else {
+                        header("Location: ../index.php?controller=adlost&action=update&id=$id&error=0");
+                    }
+
+                } else {
+                    $user_fk = $userManager->getUser($user_fk);
+                    if ($user_fk->getId()) {
+                        $ad = new AdLost($id, $animal, $name, $sex, $size, $fur, $color, $dress, $race, $number, $description, $date_lost, $date, $city, $picture, $user_fk);
+                        $adlostManager->update($ad);
+                        header("Location: ../index.php?controller=adlost&action=view&success=1");
+                    }
                 }
             }
+            $this->return("update/updateLostView", "Anim'Nord : Modifier une annonce");
         }
-        $this->return("update/updateLostView", "Anim'Nord : Modifier une annonce");
     }
 
     /**
@@ -198,23 +192,25 @@ class AdLostController {
      * @param $ad
      */
     public function deleteAd($ad) {
-        if (isset($ad['id'], $ad['user_fk'], $ad['picture'])) {
-            $userManager = new UserManager();
-            $adlostManager = new AdLostManager();
+        if (isset($_SESSION["id"])) {
+            if (isset($ad['id'], $ad['user_fk'], $ad['picture'])) {
+                $userManager = new UserManager();
+                $adlostManager = new AdLostManager();
 
-            $id = intval($ad['id']);
-            $user_fk = intval($ad['user_fk']);
-            $picture = htmlentities($ad['picture']);
+                $id = intval($ad['id']);
+                $user_fk = intval($ad['user_fk']);
+                $picture = htmlentities($ad['picture']);
 
-            unlink("./assets/img/adLost/" . $picture);
+                unlink("./assets/img/adLost/" . $picture);
 
-            $user_fk = $userManager->getUser($user_fk);
-            if ($user_fk->getId()) {
-                $adlost = new AdLost($id);
-                $adlostManager->delete($adlost);
-                header("Location: ../index.php?controller=adlost&action=view&success=2");
+                $user_fk = $userManager->getUser($user_fk);
+                if ($user_fk->getId()) {
+                    $adlost = new AdLost($id);
+                    $adlostManager->delete($adlost);
+                    header("Location: ../index.php?controller=adlost&action=view&success=2");
+                }
             }
+            $this->return("delete/deleteLostView", "Anim'Nord : Supprimer une annonce");
         }
-        $this->return("delete/deleteLostView", "Anim'Nord : Supprimer une annonce");
     }
 }
